Improved the efficiency of the raw Lisp definition of state-global-let* in many cases, and reduced code size.
Modified the check in acl2-check.lisp related to
*common-lisp-specials-and-constants*, so that it no longer causes an
error when attempting to build with ECL (which bound the symbol
COMMON-LISP:FUNCTION).

Improved the comment for with-live-state.  Also eliminated a useless
check in that definition.  For example, the following form did not
cause an error in raw Lisp (and still doesn't).

  (let ((state 17)) (with-live-state state))

Replaced lisp::system by equivalent symbol si::system in #+akcl case
of system-call, anticipating possible future ECL support, where we
would use this same akcl code for system-call.