Fixed a performance bug in raw-ev-fncall introduced in the congruent stobjs work (svn 609); the story, below, is a bit subtle.
It turns out that there has been a bug in raw-ev-fncall for some time.
The bug was kind of small in the sense that it may well not have
affected any results.  The bug was that raw-ev-fncall always used the
installed ACL2 world, (w state), even when it was called from within a
call made by ev-fncall-rec on a different world.

In passing, I noticed and fixed that bug.  But it turned out that in
some cases, the installed world is slightly longer than the world
being passed to raw-ev-fncall.  So then getprop could be (and
apparently was) really slow!

I plan to provide a proper fix for raw-ev-fncall soon.  But in the
meantime, it's no more broken than it's been for many months, but at
least ACL2 is no longer slowed down by recent changes.