Change serialize so that it no longer tries to canonicalize strings.
This fixes a bug that could cause lead to a hard error during the restoration
of fast alists, when a normed string had been canonicalized into an unnormed
string.

It should probably also slightly speed up serialize-write in most real-world
cases, at some cost in file size.  (Matt: it might be interesting to time a
regression and see if this makes any difference, but I sort of doubt it will
be significant)