With David R. and J, enhanced some comments for ACL2(p).
Fixed with-ubt! so that we don't leave a deflabel in the world when
the underlying form fails.  Thanks to Shilpi Goel for pointing out an
inconvenience with the initial implementation of disassemble$ that led
to this fix.

Name change: reverse-into-array ==> reverse-into-dmr-array.

Fixed ill-guarded calls of eq and union-eq in non-recursive-fnnames
and non-recursive-fnnames-lst, respectively.

Extended a comment as follows in note-4-4; only sublis-var and
sublis-var-lst were there before.

; To support guard verification (in new books distributed in books/system/):
; Updated several guards and termination conditions (to endp).  Functions
; affected include sublis-var and sublis-var-lst and some ancestors; subst-var
; and subst-var-lst; and subst-expr1 and subst-expr1-lst and subst-expr.
; Thanks to David Rager for his part in this effort, including his addition of
; books that verify guards for these functions.

Quoting :doc note-4-4:

| Improved dmr to show the function being called in the case of
| explicit evaluation: ``(EV-FNCALL function-being-called)''.