Improved :doc defabsstobj a bit, to clarify printing of missing events.
Improved :doc skip-proofs, linking to ld-skip-proofsp.

With David Rager, fixed a bug causing CCL-based ACL2(p) to break when
different threads attempted to print to the same output channel.

With David Rager, fixed a bug causing ACL2(p) to print goal names when
such printing was supposed to be suppressed (which was happening when
those goal names were associated with hints).

Fixed a bug in trace$ (actually, fixed supporting functions
cltl-def-from-name1 and trace$-fn-general) to avoid ill-advised
attempts to trace `functions' introduced by defabsstobj, which are
macros under the hood.  (I suppose we could trace just their *1*
functions, but we'll just deal with this case in complete analogy to
how we dealt with functions introduced by defstobj with :inline t.)

Fixed a bug in source function bogus-exported-compliants.  Quoting
:doc note-4-4:

| Fixed a bug that could cause hard Lisp errors in an encapsulate event.
| Thanks to Sol Swords for sending an example that exhibited this bug.  Here is
| a simpler such example; the bug was in how it was checked whether the
| guard for a guard-verified function (here, g) depends on some
| function introduced in the signature of the encapsulate (here, the
| function f).
|
| (encapsulate
|  ((f (x) t))
|  (local (defun f (x) (declare (xargs :guard t)) x))
|  (defun g (x)
|    (declare (xargs :guard (if (integerp x) (f x) t)))
|    x))