Fixed a problem that caused certification to fail in Lispworks.  Below are some words I've written that explain the problem.
.....

There are times where it may be tempting to redefine a function in raw
Lisp by using progn!, as follows.

(defun some-fn (...)
  ...)

(progn!
 (set-raw-mode t)
 (defun some-fn (...) ; new definition
   ...))

However, this can cause a problem in Lisps that do not compile on the
fly.  (As of this writing, this problem thus pertains to all Lisps
other than CCL and SBCL.)  Consider for example the ACL2 community
book books/serialize/unsound-read.lisp, which concludes with the
following form.

(include-raw "unsound-read-raw.lsp")

Originally the last form was instead as follows.

(progn!
 (set-raw-mode t)
 (defun unsound-read-fn (filename hons-mode verbosep)
   (let ((*ser-verbose* verbosep))
     (with-open-file (stream filename :direction :input)
       (ser-decode-from-stream t hons-mode stream)))))

But then the book books/serialize/serialize-tests.lisp failed to
certify for a LispWorks build of ACL2.  The problem was that the form
(comp t) occurs in serialize-tests.lisp before unsound-read-fn is
called.  In LispWorks, unsound-read-fn was not compiled when redefined
in raw Lisp as shown above, and hence (comp t) fetched the ACL2 loop
definition of unsound-read-fn and compiled it -- thus smashing the
preferred raw Lisp definition of this function!  An initial attempt to
fix this problem was to add the form (compile 'unsound-read-fn) after
the above defun, still inside the progn!.  But LispWorks caused an
error, complaining about a lexical closure.