Added links to ACL2-Doc, thanks to feedback from J.
Updated documentation for ACL2-Doc browser, based on change of
  control-x a a
to
  control-t g
and the addition of these (as explained in :doc acl2-doc):
  control-t G
  control-t control-t G
I've eliminated using the "control-x a" prefix because it's "owned" by
the "abbrev" mechanism, and I'm concerned that some day Emacs will
take over "control-x a a".  For now, control-x a a prints a helpful
error message, but I plan to remove this key binding after the next
release.